compiler: Allow stack allocation for mutable vars
Mutable variables use a hidden record type to implement mutability. This
record has a single field which contains the current value of the
mutable variable.

We cannot support general stack allocation of records yet as any records
with out-of-line storage require a destructor to free the storage area.
Additionally, inheriting from a record type causes the all of the record
types in the inheritance tree to be out-of-line allocated.

However, because the hidden record type only contains a single field and
cannot be inherited from we can always allocate it on the stack if it's
not captured (e.g. by a lambda enclosing its value).

In addition to reducing GC pressure this can significantly reduce
assembler size. We no longer need to store the mutable cell to memory at
all; LLVM can usually extract the single value the mutable record is
tracking.

This also tweaks the code generation for InitPair to not GC root the
pair if it is stack allocated and neither the car or cdr are GC roots.