runtime: Don't build temp results for iterators
(for-each), (string-for-each) and (vector-for-each) were basically
implemented as their mapping versions that discarded their results. This
is wasteful the temporary result was allocated and built before being
eventually discarded.

Fix by passing in a lambda to build an arbitrary container and then pass
that container in to the iterator function. For the non-mapping versions
simply use an integer 0 which should be optimised away.

This speeds up `(for-each)` on a large list by approximately 30%