compiler: Remove extra type check for proc unions
When converting a value to a procedure union if the value is a procedure
we were checking its procedureness twice. We constructed an intermediate
value with the correct type to avoid the redundant test; however, it was
not used for the subsequent conversion.

Fix by using the intermediate value we just created.