compiler: Generalise record-like stack allocation
With the latest inline storage changes for record-likes we can now stack
allocate in more situations. Generalise the existing code that only
looked for mutables.

This should improve the performance of boxes, mutable pairs and
promises in some situations.

Currently calling a procedure cell always captures it so it's hard to
trigger the stack allocation of procedures. However, the code has low
maintenance overhead so keep it in place for now.