compiler: Remove toTruthyPredicate
This was previously a special case to deal with the Scheme semantics of
any non-false value being truhty. However, this was never used by the
planner and added some special casing to the code.

Now vt.Predicate is treated the same as converting to a boolean cell:
it's only valid for actual boolean values.