compiler: Split InitProcedure
There are two distinct types of procedure cells we create:

1) Internal procedure cells have entry points with the fastcc calling
   convention and are always called with a named entry pointer. The
   entry point in the procedure cell is unused.

2) Adapter procedures cells use the normal C calling convention and may
   have their entry point loaded

We can optimise the #1 case by not storing the entry point and instead
use the entry point for additional closure storage. This also allows
LLVM to not have to produce an entry point with a given calling
convention for the function, providing for more inlining and
inter-procedural optimisations.