runtime: Optimise zero filled bytevectors
Scheme is not allowed to create bytevectors with uninitialised contents;
a fill value must always be specified. This means we take a hit on
filling the bytevector even if the fill values are unused.

For the specific case of zero-filled bytevectors we can use calloc(3) to
allocated zero filled memory. In turn this can use optimisations such as
depending on mmap(2)ed memory being zeroed.