compiler: Planning for dynamic (string-length)
This can build inline IR for `(string-length)`. This is important as its
implicitly used in most procedures that support string slices.

This requires a three loads and a comparison which typically generates
more inline code than calling out to the standard library. However, in
the corpus of functional tests we generate less code in total due to
other optimisation possibilities it opens it to LLVM.