compiler: Don't consider self cells a proc usage
The inlining heuristics will force inlining of procedures that only have
a single use. It would previously consider any procedure that took a
self argument as "using" itself. This is because self procedures used to
contain their entry point which would require planning the whole
procedure out-of-line.

Since `1e9e25b6` we no longer include the entry point for the self
argument of non-adapter procedures. Adjust our inlining logic to no
longer consider the self argument at all.

In the functional test corpus this is always either a win or neutral for
generated code size.