stdlib: Remove scheme/interfaces
This is a holdover from when we were attempting to comply with R7RS. The
`interfaces` directory contained the R7RS required identifiers. This was
then included in to `scheme` and optional `llambda` library.

This also removes the `(scheme read)` and `(scheme lazy)` libraries as
they've been superseded by llambda versions.