compiler: Extend occurrence typing for immutables
Currently occurrence typing can only constrain the type of a value. This
is effective for type checking and in the case of literal values it can
constrain the value itself to a constant. However, there are types such
as `<number>` that cannot support literal types where it would be useful
to propagate constants using the occurrence typing system.

- Rename `ConstrainType` to `ConstraintValue`

- Add a new `SubstituteWithConstant` conditional action

- Have `DynamicValueEqv` use `SubstituteWithConstant` when comparing
  against an immutable type