compiler: Introduce TypedNumberValue
There is a lot of complicated logic in NumberProcPlanner and
ArithmeticProcPlanner for dealing with number values of possibly known
types. Introduce TypedNumberValue which provides an abstraction on top
of IntermediateValue to make dealing with numeric types easier.

This is a no-op change except we now support performing inline
comparisons of flonums and integers