compiler: Optimise creating values from literals
CellValue.withSchemeType will automatically replace the value with a
constant if the new type is a literal. However, when initially creating
the value from TempValueToIntermediate we don't perform the same check.

This will usually be caught by LLVM but in some cases it could provide
more information our compiler frontend.