compiler: Allow inlining with indefinite arg types
Currently we fail inlining if we can't guarantee the argument values
will match at runtime. Instead, we can simply cast the argument values
to their wanted type.

This allows a lot more inlining. This has more effect on the LLVM IR
than the compiled binary. This is because LLVM will perform a lot of the
inlining for us. However, in quite a few cases this allows to avoid
allocations or gives us additional type information which LLVM cannot
infer itself.