compiler: Replace AssertPredicate with SignalError
`AssertPredicate` was a plan step that asserted a TempValue was true.
However, we're going to want to start planning more unconditional
assertions which makes assertions an awkward primitive.

Instead, add an `SignalError` step which unconditionally signals an
error. `PlanRuntimeAssert` is introduced as a new helper wraps a
`SignalError` in a `CondBranch` to emulate the previous behaviour.