compiler: Allow illegal code in dynamic branches
Enabling more aggressive inlining uncovered that inlining a Fibonacci
program caused an integer overflow in an untaken branch. This was
causing the compilation to fail.

This is incorrect as that code was not reachable at runtime. Teach
PlanCond to catch `SemanticException`s in dynamic branches and replace
them with runtime failures. If the branch is taken at runtime it will
fail with the same message it would've failed with at compile time.