compiler: Fix types for polymorphic inlining
We were using the storage location types for casting the inline
procedure arguments, not the polymorph type. This means we weren't
erroring out if the arguments would pass the upper bound but not the
specific polymorph type. This happened to work as we currently also plan
an invoke application at the same time which would catch the error.