compiler: Introduce InvokableProc
We want to move to a model where we can plan Lambda bodies on demand
like we currently do for Lambda polymorphs. This removes complexity in
`PlanApplication` which special cases things such as self-evaluating
lambdas as planning a procedure value can be very expensive.

The current definition to `ProcedureValue` presumes the procedure has
already been planned and its signature known. Fix this by moving the
public interface to an `InvokableProc` case class that can be requested
only when an actual invocation is required.

This only changes the public interface. Internally the implementations
need to be restructured to allow lazy planning.